結果表示を表示とエクスポートに分割し、状態表示を整理
- 409行あった training-menu-result.tsx から Markdown 生成・PDF/CSV出力・共有を
  menu-export-actions.tsx へ抽出し、本体は表示専任に
- 日付・負荷レベルの表示ヘルパーを lib/menu-labels.ts へ移設(履歴系との重複解消)
- テーブルを overflow-x-auto で包み、320px 幅でページ全体が横に崩れないように
- result-content: ローディングをカード型スケルトンに、Vercel KV/Blob の内部事情を
  並べたデバッグ表示をやめて Alert + 履歴/新規作成への導線に整理
- メニューID を含む console.log を削除、window.location.href 遷移を Link に置換
- 器具バッジ等の直書き色をダークモード対応のセマンティックトークンへ
