+++ components/result-content.tsx
@@ -3,8 +3,11 @@
 import React from "react";
 import { useSearchParams } from 'next/navigation';
 import { Button } from "@/components/ui/button";
-import { ArrowLeft } from "lucide-react";
+import { ArrowLeft, AlertCircle, SearchX } from "lucide-react";
 import Link from "next/link";
+import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
+import { Card, CardContent, CardHeader } from "@/components/ui/card";
+import { Skeleton } from "@/components/ui/skeleton";
 import TrainingMenuResult from "@/components/training-menu-result"
 
 export default function ResultContent() {
@@ -23,17 +26,15 @@ export default function ResultContent() {
 
     async function fetchData() {
       try {
-        console.log(`Fetching menu data for ID: ${menuId}`);
         const response = await fetch(`/api/get-menu?id=${menuId}`);
-        
+
         if (!response.ok) {
           const errorData = await response.json().catch(() => ({}));
           console.error("API Error Response:", errorData);
-          throw new Error(errorData.error || "Failed to fetch menu data");
+          throw new Error(errorData.error || "メニューの取得に失敗しました");
         }
-        
+
         const data = await response.json();
-        console.log("Received menu data:", data);
         setMenuData(data);
         setLoading(false);
       } catch (error: any) {
@@ -49,8 +50,19 @@ export default function ResultContent() {
   if (loading) {
     return (
       <div className="container mx-auto px-4 py-12">
-        <div className="flex flex-col items-center justify-center text-center mb-8">
-          <h1 className="text-3xl font-bold tracking-tight mb-4">読み込み中...</h1>
+        <div role="status" aria-busy="true" className="max-w-4xl mx-auto">
+          <span className="sr-only">読み込み中...</span>
+          <Card>
+            <CardHeader className="space-y-3">
+              <Skeleton className="h-8 w-2/3" />
+              <Skeleton className="h-4 w-48" />
+            </CardHeader>
+            <CardContent className="space-y-3">
+              {[0, 1, 2, 3].map((i) => (
+                <Skeleton key={i} className="h-12 w-full" />
+              ))}
+            </CardContent>
+          </Card>
         </div>
       </div>
     );
@@ -59,23 +71,22 @@ export default function ResultContent() {
   if (error) {
     return (
       <div className="container mx-auto px-4 py-12">
-        <div className="flex flex-col items-center justify-center text-center mb-8">
-          <h1 className="text-3xl font-bold tracking-tight mb-4">エラー</h1>
-          <p className="text-lg text-muted-foreground max-w-3xl mb-4">
-            メニューの取得に失敗しました: {error.message}
-          </p>
-          <div className="bg-red-50 border border-red-200 rounded-md p-4 text-left w-full max-w-2xl">
-            <h3 className="text-lg font-medium text-red-800 mb-2">デバッグ情報</h3>
-            <p className="text-sm text-red-700">
-              このエラーはVercel KVまたはBlobストレージの設定に問題がある可能性があります。
-              以下を確認してください:
-            </p>
-            <ul className="list-disc list-inside mt-2 text-sm text-red-700 space-y-1">
-              <li>Vercel KVが正しく設定されていること</li>
-              <li>Vercel Blobが正しく設定されていること</li>
-              <li>必要な環境変数が.envファイルに設定されていること</li>
-              <li>メニューIDが正しい形式であること（例: menu-12345678）</li>
-            </ul>
+        <div className="max-w-2xl mx-auto space-y-4">
+          <Alert variant="destructive">
+            <AlertCircle className="h-4 w-4" />
+            <AlertTitle>メニューの取得に失敗しました</AlertTitle>
+            <AlertDescription>{error.message}</AlertDescription>
+          </Alert>
+          <div className="flex flex-wrap gap-3">
+            <Button variant="outline" asChild>
+              <Link href="/history">
+                <ArrowLeft className="mr-2 h-4 w-4" aria-hidden="true" />
+                履歴に戻る
+              </Link>
+            </Button>
+            <Button asChild>
+              <Link href="/create">新しいメニューを作成</Link>
+            </Button>
           </div>
         </div>
       </div>
@@ -85,31 +96,22 @@ export default function ResultContent() {
   if (!menuData) {
     return (
       <div className="container mx-auto px-4 py-12">
-        <div className="flex flex-col items-center justify-center text-center mb-8">
-          <h1 className="text-3xl font-bold tracking-tight mb-4">メニューが見つかりません</h1>
-          <p className="text-lg text-muted-foreground max-w-3xl mb-4">
-            指定されたIDのメニューが見つかりませんでした。
+        <div
+          role="status"
+          className="max-w-2xl mx-auto rounded-lg border border-dashed px-6 py-12 text-center"
+        >
+          <SearchX className="mx-auto h-12 w-12 text-muted-foreground" aria-hidden="true" />
+          <p className="mt-4 text-lg font-medium">メニューが見つかりません</p>
+          <p className="mt-1 text-sm text-muted-foreground">
+            指定されたIDのメニューが見つかりませんでした。削除されたか、URLが正しくない可能性があります。
           </p>
-          <div className="bg-yellow-50 border border-yellow-200 rounded-md p-4 text-left w-full max-w-2xl">
-            <h3 className="text-lg font-medium text-yellow-800 mb-2">考えられる原因</h3>
-            <ul className="list-disc list-inside mt-2 text-sm text-yellow-700 space-y-1">
-              <li>メニューIDが無効または存在しない</li>
-              <li>ストレージから正しいインデックスデータを取得できない</li>
-              <li>Blobストレージからメニューデータを取得できない</li>
-              <li>URLパラメータが正しくない（例: <code>?id=menu-12345678</code>の形式であるべき）</li>
-            </ul>
-            <div className="mt-4 p-3 bg-white rounded border border-yellow-200">
-              <h4 className="text-md font-medium text-yellow-800 mb-2">解決方法</h4>
-              <p className="text-sm text-yellow-700 mb-2">
-                最近のVercelストレージサービスの変更により、既存のメニューデータにアクセスできなくなった可能性があります。
-              </p>
-              <a 
-                href="/create"
-                className="inline-block mt-2 px-4 py-2 bg-blue-600 text-white font-medium text-xs leading-tight rounded shadow-md hover:bg-blue-700 hover:shadow-lg focus:bg-blue-700 focus:shadow-lg focus:outline-none focus:ring-0 active:bg-blue-800 active:shadow-lg transition duration-150 ease-in-out"
-              >
-                新しいメニューを作成する
-              </a>
-            </div>
+          <div className="mt-6 flex flex-wrap justify-center gap-3">
+            <Button variant="outline" asChild>
+              <Link href="/history">履歴を確認する</Link>
+            </Button>
+            <Button asChild>
+              <Link href="/create">新しいメニューを作成する</Link>
+            </Button>
           </div>
         </div>
       </div>
@@ -119,14 +121,14 @@ export default function ResultContent() {
   return (
     <div className="container mx-auto px-4 py-12">
       <div className="mb-6">
-        <Link href="/history">
-          <Button variant="ghost" className="flex items-center gap-2">
-            <ArrowLeft className="h-4 w-4" />
+        <Button variant="ghost" asChild>
+          <Link href="/history" className="flex items-center gap-2">
+            <ArrowLeft className="h-4 w-4" aria-hidden="true" />
             履歴に戻る
-          </Button>
-        </Link>
+          </Link>
+        </Button>
       </div>
-      
+
       <div className="flex flex-col items-center justify-center text-center mb-8">
         <h1 className="text-3xl font-bold tracking-tight mb-4">生成されたメニュー</h1>
         <p className="text-lg text-muted-foreground max-w-3xl">
